Adjust feedback button to bottom-right
