+++ src/app.scss
@@ -6,29 +6,6 @@
     background: #282c34;
     position: relative;
 
-    .feedback{
-        position: absolute;
-        bottom: 5px;
-        left: 20px;
-        z-index: 20000;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-    
-        .feedback-img{
-            width: 60px;
-            height: 60px;
-            transform: scale(0.9);
-            transition: transform 330ms ease-in;
-            cursor: pointer;
-        }
-
-        >p{
-            color: #ececec;
-        };
-    }
-
     .side-menu-open{
         width: 25%;
         height: 90vh;
@@ -130,23 +107,6 @@
             top: 0px;
             height: 100vh;
             width: 85%;
-        }
-
-        .feedback{
-            position: absolute;
-            bottom: 85px;
-            // left: 70vw;
-            
-            .feedback-img{
-                width: 50px;
-                height: 50px;
-                transform: scale(0.9);
-                transition: transform 330ms ease-in;
-            }
-    
-            >p{
-                font-size: 5px
-            };
         }
     
         .side-menu-close{
