+++ src/components/MapView/styles.scss
@@ -2,6 +2,7 @@
     width: 100%;
 }
 .containers{
+    width: calc(100% - 25px);
     display: flex;
     justify-content: space-around;
 
@@ -34,6 +35,34 @@
         border-radius: 20px;
         transform: translateX(3vh);
         word-break: break-word;
+
+        .feedback{
+            position: absolute;
+            bottom: 27px;
+            right: 55px;
+            z-index: 20000;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            border: 2px solid rgb(136,169,178);
+            border-radius: 8px;
+            background: #FFF;
+        
+            .feedback-img{
+                width: 42px;
+                height: 42px;
+                transform: scale(0.9);
+                transition: transform 330ms ease-in;
+                cursor: pointer;
+            }
+    
+            >p{
+                color: black;
+                margin: 0px 5px 3px 5px;
+                font-size: 12px;
+            };
+        }
     }
 };
